// templates/electro/types.ts
export interface Product {
  id: number;
  name: string;
  brand: string;
  price: number;
  oldPrice?: number;
  imageUrl: string; // Gardé pour rétrocompatibilité (sera l'image principale)
  images: string[]; // Nouveau champ pour la galerie
  discount?: number;
  category: string;
  promo?: boolean;
  material?: string;
  description?: string;
  quantity: number;
  specifications?: { name: string; value: string; }[];
}